refactor(api): deprecate route POST /schooling-registration-dependent-users/generate-username-password
in favor of POST /sco-organization-learners/username-password-generation
