+++ api/lib/application/organization-learner-dependent-users/index.js
@@ -1,8 +1,5 @@
 const Joi = require('joi').extend(require('@joi/date'));
 
-const securityPreHandlers = require('../security-pre-handlers');
-const identifiersType = require('../../domain/types/identifiers-type');
-
 const inePattern = new RegExp('^[0-9]{9}[a-zA-Z]{2}$');
 const inaPattern = new RegExp('^[0-9]{10}[a-zA-Z]{1}$');
 
@@ -10,37 +7,6 @@ const organizationLearnerDependentUserController = require('./organization-learn
 
 exports.register = async function (server) {
   server.route([
-    {
-      method: 'POST',
-      path: '/api/schooling-registration-dependent-users/generate-username-password',
-      config: {
-        pre: [
-          {
-            method: securityPreHandlers.checkUserBelongsToScoOrganizationAndManagesStudents,
-            assign: 'belongsToScoOrganizationAndManageStudents',
-          },
-        ],
-        handler: organizationLearnerDependentUserController.generateUsernameWithTemporaryPassword,
-        validate: {
-          options: {
-            allowUnknown: true,
-          },
-          payload: Joi.object({
-            data: {
-              attributes: {
-                'organization-id': identifiersType.organizationId,
-                'schooling-registration-id': identifiersType.schoolingRegistrationId,
-              },
-            },
-          }),
-        },
-        notes: [
-          "- Génère un identifiant pour l'élève avec un mot de passe temporaire \n" +
-            "- La demande de génération d'identifiant doit être effectuée par un membre de l'organisation à laquelle appartient l'élève.",
-        ],
-        tags: ['api', 'organizationLearnerDependentUser', 'username'],
-      },
-    },
     {
       method: 'POST',
       path: '/api/schooling-registration-dependent-users/recover-account',
