+++ api/lib/application/sco-organization-learners/sco-organization-learner-controller.js
@@ -2,6 +2,7 @@ const usecases = require('../../domain/usecases');
 const scoOrganizationLearnerSerializer = require('../../infrastructure/serializers/jsonapi/sco-organization-learner-serializer');
 const organizationLearnerUserAssociationSerializer = require('../../infrastructure/serializers/jsonapi/organization-learner-user-association-serializer');
 const { extractLocaleFromRequest } = require('../../infrastructure/utils/request-response-utils');
+const organizationLearnerDependentUserSerializer = require('../../infrastructure/serializers/jsonapi/organization-learner-dependent-user-serializer');
 
 module.exports = {
   async reconcileScoOrganizationLearnerManually(request, h) {
@@ -190,5 +191,28 @@ module.exports = {
     return h
       .response(scoOrganizationLearnerSerializer.serializeCredentialsForDependent(scoOrganizationLearner))
       .code(200);
+  },
+
+  async generateUsernameWithTemporaryPassword(request, h) {
+    const payload = request.payload.data.attributes;
+    const organizationId = payload['organization-id'];
+    const organizationLearnerId = payload['schooling-registration-id'] || payload['organization-learner-id'];
+
+    const result = await usecases.generateUsernameWithTemporaryPassword({
+      organizationLearnerId,
+      organizationId,
+    });
+
+    if (h.request.path === '/api/schooling-registration-dependent-users/generate-username-password') {
+      const organizationLearnerWithGeneratedUsernamePasswordResponse =
+        organizationLearnerDependentUserSerializer.serialize(result);
+      return h
+        .response(organizationLearnerWithGeneratedUsernamePasswordResponse)
+        .code(200)
+        .header('Deprecation', 'true')
+        .header('Link', '/api/sco-organization-learners/username-password-generation; rel="successor-version"');
+    }
+
+    return h.response(scoOrganizationLearnerSerializer.serializeCredentialsForDependent(result)).code(200);
   },
 };
