+++ api/lib/infrastructure/serializers/jsonapi/sco-organization-learner-serializer.js
@@ -21,7 +21,7 @@ module.exports = {
 
   serializeCredentialsForDependent(scoOrganizationLearner) {
     return new Serializer('dependent-users', {
-      attributes: ['generatedPassword'],
+      attributes: ['username', 'generatedPassword'],
     }).serialize(scoOrganizationLearner);
   },
 };
