+++ api/tests/acceptance/application/organization-learner-dependent-user-controller_test.js
@@ -1,4 +1,4 @@
-const { databaseBuilder, expect, knex, generateValidRequestAuthorizationHeader } = require('../../test-helper');
+const { databaseBuilder, expect, knex } = require('../../test-helper');
 
 const createServer = require('../../../server');
 
@@ -11,109 +11,6 @@ describe('Acceptance | Route | organization-learner-dependent-user', function ()
 
   afterEach(async function () {
     await knex('authentication-methods').delete();
-  });
-
-  describe('POST /api/schooling-registration-dependent-users/generate-username-password', function () {
-    let organizationId;
-    let organizationLearnerId;
-    let options;
-
-    beforeEach(async function () {
-      organizationId = databaseBuilder.factory.buildOrganization({
-        type: 'SCO',
-        isManagingStudents: true,
-      }).id;
-      const userId = databaseBuilder.factory.buildUser.withRawPassword({
-        username: null,
-      }).id;
-      databaseBuilder.factory.buildMembership({ organizationId, userId });
-      organizationLearnerId = databaseBuilder.factory.buildOrganizationLearner({
-        organizationId,
-        userId,
-      }).id;
-
-      await databaseBuilder.commit();
-
-      options = {
-        method: 'POST',
-        url: '/api/schooling-registration-dependent-users/generate-username-password',
-        headers: { authorization: generateValidRequestAuthorizationHeader(userId) },
-        payload: {
-          data: {
-            attributes: {
-              'organization-id': organizationId,
-              'schooling-registration-id': organizationLearnerId,
-            },
-          },
-        },
-      };
-    });
-
-    it('should return a 200 status after having successfully generated username and temporary password', async function () {
-      // when
-      const response = await server.inject(options);
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
-
-    it('should return a 404 status when organizationLearner does not exist', async function () {
-      // given
-      options.payload.data.attributes['schooling-registration-id'] = organizationLearnerId + 1;
-
-      // when
-      const response = await server.inject(options);
-
-      // then
-      expect(response.statusCode).to.equal(404);
-    });
-
-    it("should return a 404 status when organizationLearner's userId does not exist", async function () {
-      // given
-      const organizationLearnerId = databaseBuilder.factory.buildOrganizationLearner({
-        organizationId,
-        userId: null,
-      }).id;
-      options.payload.data.attributes['schooling-registration-id'] = organizationLearnerId;
-
-      await databaseBuilder.commit();
-
-      // when
-      const response = await server.inject(options);
-
-      // then
-      expect(response.statusCode).to.equal(404);
-    });
-
-    it('should return a 403 status when student does not belong to the same organization as organizationLearner', async function () {
-      // given
-      options.payload.data.attributes['organization-id'] = organizationId + 1;
-      options.payload.data.attributes['schooling-registration-id'] = organizationLearnerId;
-
-      // when
-      const response = await server.inject(options);
-
-      // then
-      expect(response.statusCode).to.equal(403);
-    });
-
-    it('should return a 403 status when user does not belong to the same organization as organizationLearner', async function () {
-      // given
-      const wrongOrganization = databaseBuilder.factory.buildOrganization();
-      const organizationLearnerWithWrongOrganization = databaseBuilder.factory.buildOrganizationLearner({
-        organizationId: wrongOrganization.id,
-      });
-      await databaseBuilder.commit();
-
-      options.payload.data.attributes['schooling-registration-id'] = organizationLearnerWithWrongOrganization.id;
-      options.payload.data.attributes['organization-id'] = wrongOrganization.id;
-
-      // when
-      const response = await server.inject(options);
-
-      // then
-      expect(response.statusCode).to.equal(403);
-    });
   });
 
   describe('POST /api/schooling-registration-dependent-users/recover-account', function () {
