+++ api/tests/integration/application/sco-organization-learners/sco-organization-learner-controller_test.js
@@ -7,6 +7,7 @@ const securityPreHandlers = require('../../../../lib/application/security-pre-ha
 const {
   NotFoundError,
   UserNotAuthorizedToUpdatePasswordError,
+  UserNotAuthorizedToGenerateUsernamePasswordError,
 } = require('../../../../lib/domain/errors');
 
 describe('Integration | Application | sco-organization-learners | sco-organization-learner-controller', function () {
@@ -17,6 +18,7 @@ describe('Integration | Application | sco-organization-learners | sco-organizati
     sandbox = sinon.createSandbox();
     sandbox.stub(usecases, 'createAndReconcileUserToOrganizationLearner').rejects(new Error('not expected error'));
     sandbox.stub(usecases, 'updateOrganizationLearnerDependentUserPassword').rejects(new Error('not expected error'));
+    sandbox.stub(usecases, 'generateUsernameWithTemporaryPassword').rejects(new Error('not expected error'));
     sandbox.stub(securityPreHandlers, 'checkUserBelongsToScoOrganizationAndManagesStudents');
     httpTestServer = new HttpTestServer();
     await httpTestServer.register(moduleUnderTest);
@@ -303,6 +305,126 @@ describe('Integration | Application | sco-organization-learners | sco-organizati
           const response = await httpTestServer.request(
             'POST',
             '/api/sco-organization-learners/password-update',
+            payload,
+            auth
+          );
+
+          // then
+          expect(response.statusCode).to.equal(403);
+        });
+      });
+    });
+  });
+
+  describe('#generateUsernameWithTemporaryPassword for api/schooling-registration-dependent-users/generate-username-password', function () {
+    const payload = { data: { attributes: {} } };
+    const auth = { credentials: {}, strategy: {} };
+    const generatedPassword = 'Passw0rd';
+    const username = 'john.harry0207';
+
+    beforeEach(function () {
+      securityPreHandlers.checkUserBelongsToScoOrganizationAndManagesStudents.callsFake((request, h) =>
+        h.response(true)
+      );
+      payload.data.attributes = {
+        'schooling-registration-id': 1,
+        'organization-id': 3,
+      };
+      auth.credentials.userId = domainBuilder.buildUser().id;
+    });
+
+    context('Success cases', function () {
+      it('should return an HTTP response with status code 200', async function () {
+        // given
+        usecases.generateUsernameWithTemporaryPassword.resolves({ username, generatedPassword });
+
+        // when
+        const response = await httpTestServer.request(
+          'POST',
+          '/api/schooling-registration-dependent-users/generate-username-password',
+          payload,
+          auth
+        );
+
+        // then
+        expect(response.statusCode).to.equal(200);
+        expect(response.result.data.attributes['username']).to.equal(username);
+        expect(response.result.data.attributes['generated-password']).to.equal(generatedPassword);
+      });
+    });
+
+    context('Error cases', function () {
+      context('when the student has not access to the organization an error is thrown', function () {
+        it('should resolve a 403 HTTP response', async function () {
+          // given
+          usecases.generateUsernameWithTemporaryPassword.rejects(
+            new UserNotAuthorizedToGenerateUsernamePasswordError()
+          );
+
+          // when
+          const response = await httpTestServer.request(
+            'POST',
+            '/api/schooling-registration-dependent-users/generate-username-password',
+            payload,
+            auth
+          );
+
+          // then
+          expect(response.statusCode).to.equal(403);
+        });
+      });
+    });
+  });
+
+  describe('#generateUsernameWithTemporaryPassword for api/sco-organization-learners/username-password-generation', function () {
+    const payload = { data: { attributes: {} } };
+    const auth = { credentials: {}, strategy: {} };
+    const generatedPassword = 'Passw0rd';
+    const username = 'john.harry0207';
+
+    beforeEach(function () {
+      securityPreHandlers.checkUserBelongsToScoOrganizationAndManagesStudents.callsFake((request, h) =>
+        h.response(true)
+      );
+      payload.data.attributes = {
+        'organization-learner-id': 1,
+        'organization-id': 3,
+      };
+      auth.credentials.userId = domainBuilder.buildUser().id;
+    });
+
+    context('Success cases', function () {
+      it('should return an HTTP response with status code 200', async function () {
+        // given
+        usecases.generateUsernameWithTemporaryPassword.resolves({ username, generatedPassword });
+
+        // when
+        const response = await httpTestServer.request(
+          'POST',
+          '/api/sco-organization-learners/username-password-generation',
+          payload,
+          auth
+        );
+
+        // then
+        expect(response.statusCode).to.equal(200);
+        expect(response.result.data.attributes['username']).to.equal(username);
+        expect(response.result.data.attributes['generated-password']).to.equal(generatedPassword);
+      });
+    });
+
+    context('Error cases', function () {
+      context('when the student has not access to the organization an error is thrown', function () {
+        it('should resolve a 403 HTTP response', async function () {
+          // given
+          usecases.generateUsernameWithTemporaryPassword.rejects(
+            new UserNotAuthorizedToGenerateUsernamePasswordError()
+          );
+
+          // when
+          const response = await httpTestServer.request(
+            'POST',
+            '/api/sco-organization-learners/username-password-generation',
             payload,
             auth
           );
