+++ api/tests/unit/application/sco-organization-learners/sco-organization-learner-controller_test.js
@@ -5,6 +5,7 @@ const usecases = require('../../../../lib/domain/usecases');
 const scoOrganizationLearnerController = require('../../../../lib/application/sco-organization-learners/sco-organization-learner-controller');
 const scoOrganizationLearnerSerializer = require('../../../../lib/infrastructure/serializers/jsonapi/sco-organization-learner-serializer');
 const organizationLearnerUserAssociationSerializer = require('../../../../lib/infrastructure/serializers/jsonapi/organization-learner-user-association-serializer');
+const organizationLearnerDependentUserSerializer = require('../../../../lib/infrastructure/serializers/jsonapi/organization-learner-dependent-user-serializer');
 
 describe('Unit | Application | Controller | sco-organization-learner', function () {
   describe('#reconcileScoOrganizationLearnerManually', function () {
@@ -284,6 +285,49 @@ describe('Unit | Application | Controller | sco-organization-learner', function 
         path: '/api/sco-organization-learners/password-update',
       };
       const response = await scoOrganizationLearnerController.updatePassword(request, hFake);
+
+      // then
+      expect(response.headers['Deprecation']).to.not.exist;
+    });
+  });
+
+  describe('#generateUsernameWithTemporaryPassword', function () {
+    const userId = 2;
+    const request = {
+      auth: { credentials: { userId } },
+      payload: { data: { attributes: {} } },
+    };
+
+    beforeEach(function () {
+      sinon.stub(usecases, 'generateUsernameWithTemporaryPassword');
+      usecases.generateUsernameWithTemporaryPassword.resolves();
+      sinon.stub(organizationLearnerDependentUserSerializer, 'serialize');
+      organizationLearnerDependentUserSerializer.serialize.resolves();
+      usecases.generateUsernameWithTemporaryPassword.resolves();
+      sinon.stub(scoOrganizationLearnerSerializer, 'serializeCredentialsForDependent');
+      scoOrganizationLearnerSerializer.serializeCredentialsForDependent.resolves();
+    });
+
+    it('should return information about deprecation when old route is used', async function () {
+      // when
+      hFake.request = {
+        path: '/api/schooling-registration-dependent-users/generate-username-password',
+      };
+      const response = await scoOrganizationLearnerController.generateUsernameWithTemporaryPassword(request, hFake);
+
+      // then
+      expect(response.headers['Deprecation']).to.equal('true');
+      expect(response.headers['Link']).to.equal(
+        '/api/sco-organization-learners/username-password-generation; rel="successor-version"'
+      );
+    });
+
+    it('should not return information about deprecation when new route is used', async function () {
+      // when
+      hFake.request = {
+        path: '/api/sco-organization-learners/username-password-generation',
+      };
+      const response = await scoOrganizationLearnerController.generateUsernameWithTemporaryPassword(request, hFake);
 
       // then
       expect(response.headers['Deprecation']).to.not.exist;
